% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Calculating basic descriptive statistics}
\usage{
desc(data, vars)
}
\arguments{
\item{data}{The data frame}

\item{vars}{The vector of columns/variables of interest}
}
\value{
The tibble with basic descriptive statistics of selected columns
}
\description{
Created to provide basic descriptive statistics in the tibble format
}
\examples{
data("iris")
vars_select <- c("Sepal.Length", "Sepal.Width")
desc(iris, vars_select) 
}
