% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descby.R
\name{descby}
\alias{descby}
\title{Calculating group descriptive statistics}
\usage{
descby(data, vars, by)
}
\arguments{
\item{data}{The data frame}

\item{vars}{The vector of numeric columns/variables of interest}

\item{by}{The vector of factor columns/variables representing groups}
}
\value{
The tibble with group descriptive statistics of selected columns
}
\description{
Created to provide group descriptive statistics in the tibble format
}
\examples{
data("iris")
vars_select <- c("Sepal.Length", "Sepal.Width")
group_vars <- c("Species")
descby(iris, vars_select, group_vars)
}
