% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Calculating basic descriptive statistics}
\usage{
describe(data, varnames)
}
\arguments{
\item{data}{The data frame}

\item{varnames}{The vector of numeric columns/variables of interest}
}
\value{
The tibble with basic descriptive statistics of selected columns
}
\description{
Created to provide basic descriptive statistics in the tibble format
}
\examples{
data("iris")
vars_select <- c("Sepal.Length", "Sepal.Width")
describe(iris, vars_select) 
}
