% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_panel.R
\name{describe_panel}
\alias{describe_panel}
\title{Calculating descriptive statistics for selected variables in the panel data}
\usage{
describe_panel(data, varnames, entity)
}
\arguments{
\item{data}{The data frame}

\item{varnames}{The vector of numeric variables of interest}

\item{entity}{The object id variable}
}
\value{
The tibble with the mean, the variation, and the standard deviation of selected variables for panel data.
}
\description{
Inspired by Stata's xtsum command and corresponding discussion at https://stackoverflow.com/questions/49282083/xtsum-command-for-r. 
Calculates the mean, the variance, and the standard deviation of selected variables for panel data. 
Decomposes the variance and the standard deviation into between and within components.
}
\examples{
data("Crime", package = "plm")
vars_select <- c("crmrte", "polpc")
describe_panel(Crime, vars_select, "county")
}
