% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_panel_1.R
\name{describe_panel_1}
\alias{describe_panel_1}
\title{Calculating descriptive statistics for one variable in the panel data}
\usage{
describe_panel_1(data, varname, entity)
}
\arguments{
\item{data}{The data frame}

\item{varname}{The numeric variable of interest}

\item{entity}{The object id variable}
}
\value{
The tibble with the mean, the variation, and the standard deviation of selected variable for panel data.
}
\description{
Inspired by Stata's xtsum command and corresponding discussion at https://stackoverflow.com/questions/49282083/xtsum-command-for-r. 
Calculates the mean, the variance, and the standard deviation of selected variable for panel data. 
Decomposes the variance and the standard deviation into between and within components.
}
\examples{
data("Crime", package = "plm")
describe_panel_1(Crime, crmrte, county)
}
