% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_incomplete.R
\name{explore_incomplete}
\alias{explore_incomplete}
\title{Providing a table of entities with incomplete observations (missing values)}
\usage{
explore_incomplete(data, idvar)
}
\arguments{
\item{data}{The data frame}

\item{idvar}{Entities' identifier}
}
\value{
The tibble containing entities with number of missings for each variable, as well as total number of missings. The tibble is arranged by total number of missings.
}
\description{
Can be helpful to detect if panel dataset is balanced or unbalanced.
}
\examples{
# take balanced data and make it unbalanced
# by deletion of 2nd row (2nd time period for first individual)
data("Grunfeld", package = "plm")
Grunfeld_missing_period <- Grunfeld[-2, ]
detect_incomplete(Grunfeld_missing_period, firm)
}
