% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_incomlete.R
\name{list_incomplete}
\alias{list_incomplete}
\title{Providing a table of entities with incomplete observations (missing values)}
\usage{
list_incomplete(data, idvar)
}
\arguments{
\item{data}{The data frame}

\item{idvar}{Entities' identifier}
}
\value{
The vector containing entities with missing values
# take balanced data and make it unbalanced
# by deletion of 2nd row (2nd time period for first individual)
data("Grunfeld", package = "plm")
Grunfeld_missing_period <- Grunfeld[-2, ]
list_incomplete(Grunfeld_missing_period, firm)
}
\description{
Can be helpful to detect if panel dataset is balanced or unbalanced.
}
