% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{number}
\alias{number}
\title{Counting the number of non-missing values}
\usage{
number(varname, na.rm = TRUE)
}
\arguments{
\item{varname}{The column/variable of interest}
}
\value{
The number of non-missing values in x
}
\description{
Created to count the number of non-missing values in a column inside dplyr::summarise()
}
\examples{
library(dplyr);
data("cars");
cars \%>\% summarise(across(everything(), list(n = number, mean = mean, sd = sd), na.rm = TRUE));
cars \%>\% summarise(across(everything(), \(x) number(x, na.rm = TRUE)));
}
