% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heterog.R
\name{plot_heterog}
\alias{plot_heterog}
\title{Plotting heterogeneity among groups}
\usage{
plot_heterog(data, varname, by)
}
\arguments{
\item{data}{The data frame in long panel format or tidy format}

\item{varname}{The numeric variable of interest}

\item{by}{The grouping variable}
}
\value{
The ggplot object
}
\description{
Created to provide insights about heterogeneity among groups using `ggplot2` functions. Typically for panel data, grouping by object identifiers or periods is assumed.
}
\examples{
data("Fatalities", package = "AER")
plot_heterog(Fatalities, fatalrate, year)
plot_heterog(Fatalities, fatalrate, state)
}
