% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unbalanced.R
\name{plot_unbalanced}
\alias{plot_unbalanced}
\title{Draws a heatmap showing the structure of the unbalanced panel, i.e., for which periods there are missing observations by entity}
\usage{
plot_unbalanced(data, idvar, timevar)
}
\arguments{
\item{data}{The data frame}

\item{idvar}{Entities' identifier}

\item{timevar}{Time identifier}
}
\value{
The heatmap with missing time periods coloured red
}
\description{
Draws a heatmap showing the structure of the unbalanced panel, i.e., for which periods there are missing observations by entity
}
\examples{
# Load the data
data("USSeatBelts", package = "AER")

# Create unbalanced dataset
USSeatBelts_unbalanced <- USSeatBelts \%>\% 
  filter(!(year == 1995 & str_detect(state, "^N"))) \%>\%
  filter(!(year == 1985 & str_detect(state, "^A")))

# Draw the heatmap
plot_unbalanced(USSeatBelts_unbalanced, state, year)

}
