% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_by.R
\name{describe_by}
\alias{describe_by}
\title{Calculating group descriptive statistics}
\usage{
describe_by(data, varnames, groupvar)
}
\arguments{
\item{data}{The data frame}

\item{varnames}{The vector of numeric columns/variables of interest}

\item{groupvar}{The vector of factor columns/variables representing groups}
}
\value{
The tibble with group descriptive statistics of selected columns
}
\description{
Created to provide group descriptive statistics in the tibble format
}
\examples{
data("iris")
describe_by(iris, c(Sepal.Length, Sepal.Width), Species)
}
