% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_incomlete.R
\name{list_incomplete}
\alias{list_incomplete}
\title{Providing a table of entities with incomplete observations (missing values)}
\usage{
list_incomplete(data, idvar)
}
\arguments{
\item{data}{The data frame}

\item{idvar}{Entities' identifier}
}
\value{
The vector containing entities with missing values
}
\description{
Providing a table of entities with incomplete observations (missing values)
}
\examples{
data("USSeatBelts", package = "AER")
list_incomplete(USSeatBelts, state)
}
